implement remaining answer model functions
