+++ models/answer.js
@@ -1,8 +1,7 @@
 const db = require('../db');
 
-module.exports = {
-  getAnswersByQuestionId: async (questionId, page, count) => {
-    const queryText = `
+module.exports.getAnswersByQuestionId = async (questionId, page, count) => {
+  const queryText = `
       SELECT
         answers.id AS answer_id,
         answers.body AS body,
@@ -14,8 +13,27 @@ module.exports = {
       ORDER BY answers.date_written DESC
       LIMIT $2
       OFFSET (($3 - 1) * $2)`;
-    const queryValues = [questionId, count, page];
-    const result = await db.query(queryText, queryValues);
-    return result;
-  },
+  const queryValues = [questionId, count, page];
+  const result = await db.query(queryText, queryValues);
+  return result;
+};
+
+module.exports.addAnswerHelpfulById = async (answerId) => {
+  const queryText = `
+    UPDATE answers
+    SET helpful = helpful + 1
+    WHERE id=$1`;
+  const queryValues = [answerId];
+  const result = await db.query(queryText, queryValues);
+  return result;
+};
+
+module.exports.reportAnswerById = async (answerId) => {
+  const queryText = `
+    UPDATE answers
+    SET reported = true
+    WHERE id=$1`;
+  const queryValues = [answerId];
+  const result = await db.query(queryText, queryValues);
+  return result;
 };
